fix : fix typo on eslint ; fix react-router-dom version on package.json to resolve routes issues ; fix App.jsx with route creation ; state leverage for address data
