+++ package.json
@@ -24,7 +24,8 @@
     "axios": "^0.24.0",
     "emailjs-com": "^3.2.0",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react-refresh": "^1.3.5",
