+++ src/Components/LandingPage.jsx
@@ -1,5 +1,4 @@
-import React from 'react';
-
+import React, { useState } from 'react';
 import CurrentDate from './Date';
 import StartAstroPage from './ModuleAstro';
 import StartMeteoPage from './ModuleMeteo';
@@ -11,6 +10,23 @@ import Nav from '../components/Nav';
 import styles from './LandingPage.module.css';
 
 function LandingPage() {
+  const [destinationData, setDestinationData] = useState();
+  const [originData, setOriginData] = useState();
+
+  function handleOrigin(adress) {
+    setOriginData(adress);
+  }
+
+  function handleDestination(adress) {
+    setDestinationData(adress);
+  }
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    localStorage.setItem('origin', originData);
+    localStorage.setItem('destination', destinationData);
+  };
+
   return (
     <main className={styles.main}>
       <header className={styles.header}>
@@ -22,7 +38,7 @@ function LandingPage() {
         <CurrentDate />
       </div>
       <div className={styles.input}>
-        <AdressSearch />
+        <AdressSearch origin={handleOrigin} destination={handleDestination} handleSubmit={handleSubmit} />
       </div>
 
       <div className={styles.cards}>
