+++ src/components/index.jsx
@@ -1,5 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from '../App';
+import { BrowserRouter as Router } from 'react-router-dom';
 
-ReactDOM.render(<App />, document.querySelector('#root'));
+ReactDOM.render(
+  <Router>
+    <App />
+  </Router>,
+  document.querySelector('#root'),
+);
